Removes solc from tests
