+++ tests/package.json
@@ -29,11 +29,12 @@
     }
   },
   "scripts": {
-    "test-with-logs": "mocha --printlogs -r ts-node/register 'tests/**/*.ts'",
-    "test": "mocha --parallel -r ts-node/register -r fixtures/contract-fixtures.ts 'tests/**/*.ts'",
+    "test-with-logs": "mocha --printlogs -r ts-node/register 'tests/**/test-*.ts'",
+    "build-contracts": "TS_NODE_TRANSPILE_ONLY=true ts-node ./pre-build-contracts.ts",
+    "test": "TS_NODE_TRANSPILE_ONLY=true mocha --parallel -r ts-node/register 'tests/**/test-*.ts'",
     "watch": "npm-watch",
     "build": "cargo build --release",
-    "non-ci-test": "mocha -r ts-node/register 'non_ci_tests/**/*.ts'"
+    "non-ci-test": "mocha -r ts-node/register 'non_ci_tests/**/test-*.ts'"
   },
   "author": "",
   "license": "ISC",
